updated fragments with necessary api requests and apikeys

+++ app/src/main/java/com/appdesk/weatherapp/model/Coord.kt
@@ -1,6 +1,6 @@
 package com.appdesk.weatherapp.model
 
 data class Coord(
-    val lat: Double,
-    val lon: Double
+    val lat: Int,
+    val lon: Int
 )
